+++ NEWS.md
@@ -1,3 +1,5 @@
+# deeptime (development version)
+
 # deeptime 1.1.1
 deeptime now requires rlang version 1.1.0 or higher. This patch version includes several bug fixes: